requirements.txt
discord~=1.0.1
aiohttp~=3.7.3
pytz==2021.1
requests==2.25.1
Pillow==8.2.0
gTTS~=2.2.2
mcstatus==5.2.0
disputils~=0.1.3
PyYAML==5.4.1
python-dotenv==0.17.1
beautifulsoup4~=4.9.1
inflect~=5.3.0
python-Levenshtein~=0.12.2
fuzzywuzzy~=0.18.0